requirements.txt
python-dotenv
requests
scikit-learn
numpy
scipy
